refactor(validation): trait and implementation in different modules
